Removed global flag
