+++ packages/boot-fns-derive/Cargo.toml
@@ -10,9 +10,6 @@ description = "Derive macro for generating contract interfaces."
 [lib]
 proc-macro = true
 
-[features]
-outside_contract=[]
-
 [dependencies]
 quote = "1"
 proc-macro2 = "1"
